backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
openai==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
numpy==1.26.3
pandas==2.1.4
slowapi==0.1.9
websockets==12.0
psutil==5.9.8
redis==5.0.1
pywebpush==2.0.3

